requirements.txt
scp==0.15.0
wget==3.2
boto3==1.35.21
PyYAML==6.0.2
paramiko==3.5.0
